lcas/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
redis==6.4.0
rq==2.4.1
httpx==0.28.1
PyYAML==6.0.2
jinja2==3.1.6
pytest==8.4.1
